requirements (1).txt
streamlit
openai
psycopg2-binary
scikit-learn
python-dotenv
numpy
